Add stake mocks
