+++ modules/proposals/engine/Cargo.toml
@@ -84,9 +84,13 @@ git = 'https://github.com/joystream/substrate-stake-module'
 package = 'substrate-stake-module'
 rev = '0516efe9230da112bc095e28f34a3715c2e03ca8'
 
+[dev-dependencies]
+mockall = "0.6.0"
+
 [dev-dependencies.runtime-io]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
 package = 'sr-io'
 rev = '0e3001a1ad6fa3d1ba7da7342a8d0d3b3facb2f3'
+
