+++ modules/proposals/engine/src/tests/mock/mod.rs
@@ -8,8 +8,15 @@ pub use runtime_primitives::{
     BuildStorage, Perbill,
 };
 use srml_support::{impl_outer_dispatch, impl_outer_event, impl_outer_origin, parameter_types};
-use srml_support::traits::{Currency, Imbalance};
 pub use system;
+
+mod balance_restorator;
+mod proposals;
+mod stakes;
+
+use balance_restorator::*;
+pub use proposals::*;
+use stakes::*;
 
 // Workaround for https://github.com/rust-lang/rust/issues/26925 . Remove when sorted.
 #[derive(Clone, PartialEq, Eq, Debug)]
@@ -65,35 +72,6 @@ impl stake::Trait for Test {
     type StakingEventsHandler = BalanceRestoratorStakingEventsHandler;
     type StakeId = u64;
     type SlashId = u64;
-}
-
-pub struct BalanceRestoratorStakingEventsHandler;
-impl stake::StakingEventsHandler<Test> for BalanceRestoratorStakingEventsHandler {
-    fn unstaked(
-        _id: &u64,
-        _unstaked_amount: stake::BalanceOf<Test>,
-        imbalance: stake::NegativeImbalance<Test>,
-    ) -> stake::NegativeImbalance<Test> {
-
-        let default_account_id = 1;
-
-        <Test as stake::Trait>::Currency::resolve_creating(
-            &default_account_id,
-            imbalance,
-        );
-
-        stake::NegativeImbalance::<Test>::zero()
-    }
-
-    fn slashed(
-        _id: &u64,
-        _slash_id: &u64,
-        _slashed_amount: stake::BalanceOf<Test>,
-        _remaining_stake: stake::BalanceOf<Test>,
-        _imbalance: stake::NegativeImbalance<Test>,
-    ) -> stake::NegativeImbalance<Test> {
-        unreachable!();
-    }
 }
 
 impl crate::Trait for Test {
@@ -165,88 +143,15 @@ pub fn initial_test_ext() -> runtime_io::TestExternalities {
     t.into()
 }
 
+// Intercepts panic in provided function, test mock expectation and restores default behaviour
+pub(crate) fn handle_mock<F: std::panic::RefUnwindSafe + Fn()>(func: F) {
+    let panicked = panics(func);
+
+    test_expectation_and_clear_mock();
+
+    assert!(!panicked);
+}
+
 pub type ProposalsEngine = crate::Module<Test>;
 pub type System = system::Module<Test>;
 pub type Balances = balances::Module<Test>;
-
-use codec::{Decode, Encode};
-use num_enum::{IntoPrimitive, TryFromPrimitive};
-use rstd::convert::TryFrom;
-use rstd::prelude::*;
-
-use srml_support::dispatch;
-
-use crate::{ProposalCodeDecoder, ProposalExecutable};
-
-/// Defines allowed proposals types. Integer value serves as proposal_type_id.
-#[derive(Debug, Eq, PartialEq, TryFromPrimitive, IntoPrimitive)]
-#[repr(u32)]
-pub enum ProposalType {
-    /// Dummy(Text) proposal type
-    Dummy = 1,
-
-    /// Testing proposal type for faults
-    Faulty = 10001,
-}
-
-impl ProposalType {
-    fn compose_executable(
-        &self,
-        proposal_data: Vec<u8>,
-    ) -> Result<Box<dyn ProposalExecutable>, &'static str> {
-        match self {
-            ProposalType::Dummy => DummyExecutable::decode(&mut &proposal_data[..])
-                .map_err(|err| err.what())
-                .map(|obj| Box::new(obj) as Box<dyn ProposalExecutable>),
-            ProposalType::Faulty => FaultyExecutable::decode(&mut &proposal_data[..])
-                .map_err(|err| err.what())
-                .map(|obj| Box::new(obj) as Box<dyn ProposalExecutable>),
-        }
-    }
-}
-
-impl ProposalCodeDecoder<Test> for ProposalType {
-    fn decode_proposal(
-        proposal_type: u32,
-        proposal_code: Vec<u8>,
-    ) -> Result<Box<dyn ProposalExecutable>, &'static str> {
-        Self::try_from(proposal_type)
-            .map_err(|_| "Unsupported proposal type")?
-            .compose_executable(proposal_code)
-    }
-}
-
-/// Testing proposal type
-#[derive(Encode, Decode, Clone, PartialEq, Eq, Debug, Default)]
-pub struct DummyExecutable {
-    pub title: Vec<u8>,
-    pub body: Vec<u8>,
-}
-
-impl DummyExecutable {
-    pub fn proposal_type(&self) -> u32 {
-        ProposalType::Dummy.into()
-    }
-}
-
-impl ProposalExecutable for DummyExecutable {
-    fn execute(&self) -> dispatch::Result {
-        Ok(())
-    }
-}
-
-/// Faulty proposal executable code wrapper. Used for failed proposal execution tests.
-#[derive(Encode, Decode, Clone, PartialEq, Eq, Debug, Default)]
-pub struct FaultyExecutable;
-impl ProposalExecutable for FaultyExecutable {
-    fn execute(&self) -> dispatch::Result {
-        Err("ExecutionFailed")
-    }
-}
-
-impl FaultyExecutable {
-    /// Converts faulty proposal type to proposal_type_id
-    pub fn proposal_type(&self) -> u32 {
-        ProposalType::Faulty.into()
-    }
-}
